# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chat backend API",
        "version": "0.1"
    },
    "paths": {
        "/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Get a list of messages",
                "description": "Returns a list of messages with users, sorted by creation date.",
                "operationId": "2d594d8eb33b4b4f1f754b4c94a3ec47",
                "responses": {
                    "200": {
                        "description": "List of messages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MessageResource"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "ERROR_GET_MESSAGES"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Create a new message",
                "description": "Stores a new message and returns the created message with user details.",
                "operationId": "d4186c8f2dfb49846fce7bf96f5d36c2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Hello, world!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "ERROR_CREATE_MESSAGE"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create or retrieve a user",
                "description": "Finds an existing user by username or creates a new one with a random color.",
                "operationId": "10551226e9bb607517f5e51dd4137ec2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "john_doe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User retrieved or created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "ERROR_MAKE_USER"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ErrorResource": {
                "title": "ErrorResource",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "MessageResource": {
                "title": "MessageResource",
                "description": "Message resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "content": {
                        "type": "string",
                        "example": "Hello World",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-03T12:00:00Z",
                        "nullable": true
                    },
                    "user": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/UserResource"
                            }
                        ],
                        "nullable": true,
                        "type": "object"
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
            "SuccessResource": {
                "title": "SuccessResource",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Success message"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "UserResource",
                "description": "User resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "username": {
                        "type": "string",
                        "example": "john_doe",
                        "nullable": false
                    },
                    "color": {
                        "type": "string",
                        "example": "#ffffff",
                        "nullable": false
                    }
                },
                "type": "object",
                "additionalProperties": false
            }
        },
        "responses": {
            "ErrorResource": {
                "description": "Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResource"
                        }
                    }
                }
            },
            "MessageResource": {
                "description": "Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/MessageResource"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "SuccessResource": {
                "description": "Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/SuccessResource"
                        }
                    }
                }
            },
            "UserResource": {
                "description": "Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "data": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "requestBodies": {
            "MessageRequest": {
                "description": "Message request",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "content"
                            ],
                            "properties": {
                                "content": {
                                    "description": "Content",
                                    "type": "string",
                                    "example": "test message"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "UserRequest": {
                "description": "User request",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "username"
                            ],
                            "properties": {
                                "username": {
                                    "description": "Username",
                                    "type": "string",
                                    "example": "user123"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Messages",
            "description": "Messages"
        },
        {
            "name": "User",
            "description": "User"
        }
    ]
}